Update Classes
Positional Index Posting
Posting & App Utils

+++ src/cecs429/indexes/PositionalIndexPosting.java
@@ -1,24 +1,24 @@
 package cecs429.indexes;
 
-import java.util.ArrayList;
 import java.util.List;
 
 /**
  * A PositionalIndexPosting encapulates a document ID and 
  * List of positions the query appeared in the corresponding document ID
  * associated with a search query component.
+ * and Wdt
  */
 public class PositionalIndexPosting {
 	private int mDocumentId;
 	private List<Integer> positions;
-	private float wdt;
+	private double wdt;
 	
 	public PositionalIndexPosting(int documentId,List<Integer> positionList) {
 		mDocumentId = documentId;
 		positions= positionList;
 	}
 	
-	public PositionalIndexPosting(int documentId,List<Integer> positionList,float wdt) {
+	public PositionalIndexPosting(int documentId,List<Integer> positionList,double wdt) {
 		mDocumentId = documentId;
 		positions= positionList;
 		this.setWdt(wdt);
@@ -36,11 +36,11 @@ public class PositionalIndexPosting {
 		this.positions = positions;
 	}
 
-	public float getWdt() {
+	public double getWdt() {
 		return wdt;
 	}
 
-	public void setWdt(float wdt) {
+	public void setWdt(double wdt) {
 		this.wdt = wdt;
 	}
 }
